# internal/api/openapi.yaml
openapi: "3.0.3"
info:
  title: CertWatch API
  description: |
    Lightweight, self-hosted SSL/TLS certificate inventory and expiration monitoring.

    All API endpoints except `/health`, `/api/version`, and the auth endpoints require a JWT token obtained via `POST /api/auth/login`.
    Include it as `Authorization: Bearer <token>` in the request header.

    Rate limiting: auth endpoints (`/api/auth/*`) are limited to 10 requests per minute per IP.
    All other authenticated endpoints have no rate limit.

    Request body size is limited to 1 MB.
  version: "1.0.0"
  contact:
    name: CertWatch
    url: https://github.com/araujofrancisco/certwatch
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/domains:
    get:
      tags: [Domains]
      summary: List domains
      description: |
        Returns all domains with optional search and enabled-state filtering.
        Tags are included in each domain object.
      operationId: listDomains
      parameters:
        - $ref: "#/components/parameters/q"
        - $ref: "#/components/parameters/enabled"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Domain list
          content:
            application/json:
              schema:
                type: object
                properties:
                  domains:
                    type: array
                    items:
                      $ref: "#/components/schemas/Domain"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: failed to list domains

    post:
      tags: [Domains]
      summary: Create a domain
      description: |
        Adds a new domain to monitor.
        The domain is auto-scanned in a background goroutine immediately after creation.
      operationId: createDomain
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDomainRequest"
      responses:
        "201":
          description: Domain created
          content:
            application/json:
              schema:
                type: object
                properties:
                  domain:
                    $ref: "#/components/schemas/Domain"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid-domain:
                  summary: Invalid domain name
                  value:
                    error: 'invalid domain name: "bad domain!"'
                duplicate:
                  summary: Duplicate domain
                  value:
                    error: domain already exists
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/domains/import:
    post:
      tags: [Domains]
      summary: Bulk import domains
      description: |
        Import multiple domains at once. Accepts either an array of objects or plain strings.
        Always returns HTTP 200 — individual results have per-item statuses.
      operationId: importDomains
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkImportRequest"
      responses:
        "200":
          description: Import results (per-item statuses)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkAddResponse"
        "400":
          description: Bad request format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missing:
                  summary: Missing domains field
                  value:
                    error: missing domains
                bad-format:
                  summary: Invalid format
                  value:
                    error: domains must be an array of objects or strings
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/domains/{id}:
    get:
      tags: [Domains]
      summary: Get a domain by ID
      operationId: getDomain
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/domainId"
      responses:
        "200":
          description: Domain details
          content:
            application/json:
              schema:
                type: object
                properties:
                  domain:
                    $ref: "#/components/schemas/Domain"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Domain not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: domain not found

    put:
      tags: [Domains]
      summary: Update a domain
      description: |
        Updates domain name, description, group, enabled status, and tags.
        If the `tags` field is present (even an empty array), existing tags are replaced.
        If the `tags` field is omitted, existing tags are preserved.
      operationId: updateDomain
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/domainId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDomainRequest"
      responses:
        "200":
          description: Domain updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  domain:
                    $ref: "#/components/schemas/Domain"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Domain not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: domain not found

    delete:
      tags: [Domains]
      summary: Delete a domain
      description: |
        Deletes a domain and cascade-deletes associated certificates and tag associations.
      operationId: deleteDomain
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/domainId"
      responses:
        "200":
          description: Domain deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Domain not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: domain not found
  /api/domains/{id}/scan:
    post:
      tags: [Domains]
      summary: Scan a domain
      description: |
        Enqueues an immediate certificate scan across all protocols in priority order:
        HTTPS (5s) -> CT (10s). First success wins. The scan runs asynchronously
        through the background scan queue; this endpoint returns immediately.
        Poll GET /api/domains/{id}/certificates for results.
      operationId: scanDomain
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/domainId"
      responses:
        "202":
          description: Scan queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: queued
        "400":
          description: Invalid domain ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: invalid domain id
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Domain not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: domain not found
  /api/scanqueue:
    get:
      tags: [Domains]
      summary: Scan queue status
      description: |
        Returns the number of scans currently pending in the queue and in flight.
      operationId: scanQueueStatus
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Queue statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  pending:
                    type: integer
                    description: Scans waiting to be dispatched
                    example: 3
                  inflight:
                    type: integer
                    description: Scans currently being executed
                    example: 2
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from `POST /api/auth/login`.
        Format: `Authorization: Bearer <token>`
  parameters:
    domainId:
      name: id
      in: path
      required: true
      description: Domain ID
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1
    q:
      name: q
      in: query
      required: false
      description: Search query (LIKE match against domain/description for domains, subject/issuer for certificates)
      schema:
        type: string
      example: example
    enabled:
      name: enabled
      in: query
      required: false
      description: Filter by enabled state (accepts true/false/1/0)
      schema:
        type: boolean
      example: true
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: unauthorized
...
  schemas:
    ErrorResponse:
      description: Standard error response
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Domain:
      description: Monitored domain with optional tags
      type: object
      properties:
        id:
          type: integer
          format: int64
        domain:
          type: string
          description: Domain name (lowercased, trimmed)
          example: example.com
        description:
          type: string
          maxLength: 500
          description: Optional description (max 500 characters)
          example: My site
        enabled:
          type: boolean
          description: Whether monitoring is active
        group:
          type: string
          maxLength: 100
          description: Optional group name (max 100 characters)
          example: production
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - domain
        - enabled
        - created_at
        - updated_at
    Tag:
      description: Tag with random color assignment on first use
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          description: Tag name (lowercased, trimmed)
          example: tls
        color:
          type: string
          description: Hex color code (randomly assigned from 10-color palette)
          example: "#0d6efd"
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - color
        - created_at
    CreateDomainRequest:
      type: object
      properties:
        domain:
          type: string
          description: Domain name to monitor (will be lowercased and trimmed)
          example: example.com
        description:
          type: string
          maxLength: 500
          description: Optional description (max 500 characters)
          example: My site
        group:
          type: string
          maxLength: 100
          description: Optional group name (max 100 characters)
          example: production
        tags:
          type: array
          items:
            type: string
          description: Optional tag names (will be created if they don't exist)
          example: ["tls", "wildcard"]
      required:
        - domain
    UpdateDomainRequest:
      type: object
      properties:
        domain:
          type: string
          description: Updated domain name
          example: example.org
        description:
          type: string
          maxLength: 500
          description: Updated description (max 500 characters)
        group:
          type: string
          maxLength: 100
          description: Updated group name (max 100 characters)
        enabled:
          type: boolean
          description: Updated enabled state
        tags:
          type: array
          items:
            type: string
          description: |
            Replaces all existing tags if present.
            If omitted, existing tags are preserved.
          example: ["updated-tag"]
      required:
        - domain
        - enabled
    BulkImportRequest:
      type: object
      description: Accepts either an array of objects or an array of plain strings
      properties:
        domains:
          oneOf:
            - type: array
              items:
                $ref: "#/components/schemas/BulkDomainObject"
            - type: array
              items:
                type: string
              example: ["example.com", "example.org"]
      required:
        - domains
    BulkDomainObject:
      type: object
      properties:
        domain:
          type: string
          description: Domain name
          example: example.com
        description:
          type: string
          maxLength: 500
          description: Optional description
        group:
          type: string
          maxLength: 100
          description: Optional group name
        tags:
          type: array
          items:
            type: string
      required:
        - domain
    BulkAddResult:
      type: object
      properties:
        domain:
          type: string
          description: Domain name
        status:
          type: string
          enum: [created, skipped, error]
          description: Import status
        error:
          type: string
          description: Error message (only on skipped/error status)
        description:
          type: string
        group:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - domain
        - status
    BulkAddSummary:
      type: object
      properties:
        total:
          type: integer
          description: Total domains processed
        created:
          type: integer
          description: Successfully created
        skipped:
          type: integer
          description: Skipped (already exists)
        errors:
          type: integer
          description: Failed (invalid domain)
      required:
        - total
        - created
        - skipped
        - errors
    BulkAddResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/BulkAddResult"
        summary:
          $ref: "#/components/schemas/BulkAddSummary"
      required:
        - results
        - summary
...
    DeleteResponse:
      type: object
      properties:
        status:
          type: string
          example: deleted
      required:
        - status
...